+++ bin/interpreter.dart
@@ -1,0 +1,3 @@
+import 'd4rt.dart' as d4rt_cli;
+
+void main(List<String> args) => d4rt_cli.main(args);