+++ example/hello.dart
@@ -1,0 +1,6 @@
+void main(List<String> args) {
+  print('Hello from d4rt interpreter CLI!');
+  if (args.isNotEmpty) {
+    print('Arguments received: $args');
+  }
+}