+++ test/invoke_interpreted_function_test.dart
@@ -156,21 +156,8 @@ void main() {
     });
 
     test('Strong interpreted function integration documentation example', () {
-      T Function() bindFunction0<T>(D4rt d, InterpretedFunction fn) {
-        return () => d.invokeInterpretedFunction(fn, []);
-      }
-
       T Function(A) bindFunction1<T, A>(D4rt d, InterpretedFunction fn) {
         return (A a) => d.invokeInterpretedFunction(fn, [a]);
-      }
-
-      T Function(A, B) bindFunction2<T, A, B>(D4rt d, InterpretedFunction fn) {
-        return (A a, B b) => d.invokeInterpretedFunction(fn, [a, b]);
-      }
-
-      T Function(A, B, C) bindFunction3<T, A, B, C>(
-          D4rt d, InterpretedFunction fn) {
-        return (A a, B b, C c) => d.invokeInterpretedFunction(fn, [a, b, c]);
       }
 
       final source = """